requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
numpy==1.26.4
scipy==1.13.0
tensorflow==2.16.1
librosa==0.10.2
soundfile==0.12.1
Pillow==10.3.0
matplotlib==3.9.0